+++ ProjectEuler/ProjectEuler/Problems/Problem3.h
@@ -14,7 +14,7 @@ namespace Problems{
 		List<int> primes;
 	public:
 		Problem3(void);
-		int run(void);
+		__int64 run(void);
 	};
 }
 #endif
